COM-3009: rename and move component
